LlmOpenAIへ完全移行。VAD,RMSをバイトで保持。テスト強化

+++ src/test/java/audio/stt/Lfm2AudioSpeechToTextTest.java
@@ -33,7 +33,7 @@ class Lfm2AudioSpeechToTextTest {
                     "Perform ASR in japanese.",
                     Duration.ofSeconds(5)));
             AudioBuffer audio = new AudioBuffer(16_000, 1);
-            audio.append(new short[] {1, 2}, 0, new float[] {0.0f, 0.0f}, new float[] {0.0f, 0.0f});
+            audio.append(new short[] {1, 2}, 0, new byte[] {0, 0}, new byte[] {0, 0});
 
             Transcription transcription = speechToText.transcribe(audio, 0, 2, "");
